Exception Handling Exercise
Finished the exercise and am now going back to work on my math library
for the rest of the day. Need to fix my matrix multiplication 3x3 & 4x4
for I am not getting the right results for some reason. Math must be
wrong so I should just have to rewrite that and then it should be good
to go after that.