VAOs and Multiple objects
Broke down what I had from the tutorials and restarted with the focus
being on VAOs. Got them working and figured out a way to draw multiple
objects. Next week will be taking this and implementing it into the
Knight framework. After that it'll be time to mess with text